backend/requirements.txt
# MapToPoster Backend - Python Requirements

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Existing dependencies (from original requirements.txt)
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
geographic==2.1
geopandas==1.1.2
geopy==2.4.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
networkx==3.6.1
numpy==2.4.0
osmnx==2.0.7
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pyogrio==0.12.1
pyparsing==3.3.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scipy==1.16.3
shapely==2.1.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.3
urllib3==2.6.3

# Optional dependencies (for plugins)
redis==5.0.1  # Cache service
boto3==1.34.22  # S3 storage service
stripe==7.11.0  # Payment service

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing FastAPI
black==24.1.1
flake8==7.0.0
